Taking into account Apollo fragment interpolation
In Apollo, fragment interpolation is now valid within the `gql` TaggedTemplateLiteral, but only outside of top-level structures like `query` or `mutation`. In other clients like Lokka and Relay, fragment interpolation actuall results in a text substitution, but in Apollo this gets evaluated at runtime, so we only need to validate the location of the fragment interpolation.