Allow specifying the `env` option in all rules
Previously, the ESLint config schema only allowed the `env` option for
the "graphql/template-strings" rule and the "graphql/required-fields"
rule.

This commit allows the `env` option for the "graphql/named-operations"
rule and the "graphql/capitalized-type-name" rule as well, by moving the
ESLint config schema for the `env` option to the schema part shared by
all rules.

The `env` option is valuble to all rules, since it is used for the
slight parsing differences in the GraphQL syntax between apollo, relay
and lokka.