Fix crash in "graphql/required-fields"
Previously, this configuration:

    "graphql/required-fields": ["error", {}],

... resulted into this crash:

    TypeError: Cannot read property 'forEach' of undefined
        at Object.Field (/eslint-plugin-graphql/lib/rules.js:31:21)

That's because the `requiredFields` option of the
"graphql/required-fields" rule is required, but wasn't marked as such in
the ESLint config schema.

This commit marks the `requiredFields` option as required, resulting in
this error being reported by ESLint:

    Configuration for rule "graphql/required-fields" is invalid:
    Value "[object Object]" should have required property 'requiredFields'.

It would be nice to also add `minItems: 1` so that the following
configuration would be invalid as well:

    "graphql/required-fields": ["error", {requiredFields: []}],

But that would be a breaking change, so that wasn't included.