Eslint config validation (#86)
* Fix .graphqlconfig support for "graphql/capitalized-type-name"

When trying to use the "graphql/capitalized-type-name" rule with a
.graphqlconfig instead of the `schema*` options, this error was
previously reported by ESLint:

    Configuration for rule "graphql/capitalized-type-name" is invalid:
      Value "[object Object]" should have required property '.schemaJson'.
      Value "[object Object]" should have required property '.schemaJsonFilepath'.
      Value "[object Object]" should have required property '.schemaString'.
      Value "[object Object]" should match exactly one schema in oneOf.

With this commit it works as expected.

This was done by syncing the ESLint config schema for the
"graphql/capitalized-type-name" rule with the other rules.

* Allow specifying the `env` option in all rules

Previously, the ESLint config schema only allowed the `env` option for
the "graphql/template-strings" rule and the "graphql/required-fields"
rule.

This commit allows the `env` option for the "graphql/named-operations"
rule and the "graphql/capitalized-type-name" rule as well, by moving the
ESLint config schema for the `env` option to the schema part shared by
all rules.

The `env` option is valuble to all rules, since it is used for the
slight parsing differences in the GraphQL syntax between apollo, relay
and lokka.

* Fix missing rule options validation

Previously, providing no options at all to any of the rules provided by
this plugin resulted in nothing happening at all. Examples:

    "graphql/template-strings": "error",
    "graphql/required-fields": ["error"],

That's not very helpful, and clearly not intended since `minLength: 1`
was specified in the ESLint config schema for all rules. However, the
correct name is "minItems", not "minLength". (I suspect this error comes
from confusion between different JSON schema versions.)

This commit changes "minLength" to "minItems", resulting in this error
being reported by ESLint (for example):

    Configuration for rule "graphql/template-strings" is invalid:
    Value "" should NOT have less than 1 items.

* Fix crash in "graphql/required-fields"

Previously, this configuration:

    "graphql/required-fields": ["error", {}],

... resulted into this crash:

    TypeError: Cannot read property 'forEach' of undefined
        at Object.Field (/eslint-plugin-graphql/lib/rules.js:31:21)

That's because the `requiredFields` option of the
"graphql/required-fields" rule is required, but wasn't marked as such in
the ESLint config schema.

This commit marks the `requiredFields` option as required, resulting in
this error being reported by ESLint:

    Configuration for rule "graphql/required-fields" is invalid:
    Value "[object Object]" should have required property 'requiredFields'.

It would be nice to also add `minItems: 1` so that the following
configuration would be invalid as well:

    "graphql/required-fields": ["error", {requiredFields: []}],

But that would be a breaking change, so that wasn't included.

* Clarify the "graphql/required-fields" rule examples

The three examples are so similar that it is hard to see the difference
without a little bit of text guiding you.

* Improve docs regarding common options

* Update changelog

* Allow specifying no options at all for some rules

For rules which have no required options for every option group it is
valid to specify no configuration at all:

    "graphql/named-operations": "error",
    "graphql/capitalized-type-name": "error",
    "graphql/template-strings": "error",

The above is now equivalent to:

    "graphql/named-operations": ["error", {}],
    "graphql/capitalized-type-name": ["error", {}],
    "graphql/template-strings": ["error", {}],

Note that this is still a config error, because the `requiredFields`
option is required:

    "graphql/required-fields": "error",