Properly handle multi-line placeholder replacements
Previously template literal replacemenets were made by constructing
dummy strings of a particular length. Unfortunately if the inside of a
tag has whitespace or multiple lines, this fails to work.

This adjusts the implementation to keep a mapping of source code
locations to query document locations, and look up from that mapping to
correlate errors.