Set the babel target to the same minimum Node.js version via `browserslist`.
By default, if no `targets` option is passed, or `browserslist` isn't set in
the `package.json`, `@babel/preset-env` only transpiles to ES2015/ES6.

Since we're only testing Node.js 10+ and we've set the `engines` property to
indicate our intention of only supporting Node.js 10+, I think we can
transpile to a newer target safely.  Not truly important to count bytes in
this package, but this shaves off an unimpressive 8kB (but also a 20%
reduction for this package).